+++ src/Markdown/Markdown.php
@@ -11,11 +11,12 @@
  * that were distributed with this source code.
  */
 
-namespace Eventum;
+namespace Eventum\Markdown;
 
-use Eventum\CommonMark\UserMentionGenerator;
 use Eventum\Config\Paths;
+use Eventum\Event;
 use Eventum\EventDispatcher\EventManager;
+use Eventum\Markdown\CommonMark\UserMentionGenerator;
 use HTMLPurifier;
 use HTMLPurifier_HTML5Config;
 use League\CommonMark\CommonMarkConverter;
